Fixing a bug in the `khelper` + introducing some new functionalities
There was a bug in the `build` function of the `khelper` that
unwillingly disabled the `CONFIG_DEBUG_INFO` option in the kernel
configuration file. Moreover, now there is a more fine-grained
configuration for the kernel to be built (enable/disable eBPF,
syzkaller support, etc.).

There are two new scripts used in a `debootstrap`'ed context to run
the machine, and then install headers. This is required for a
convenient environment with `bpfscript`.