Introducing `bootstrap` to build a minimal Arch Linux image
Sometimes is useful to have a minimal Linux distribution with a
full-fledged package manager. However, one of the needs that I often
have is reproducibility and simplicity: relying on a common
full Linux distribution was not an option.

With `pacstrap` is quite easy to get a proper and minimal Arch Linux
system with a high level of reproducibility. This functionality,
however, is not completely developed in `khelper`, but its core
functionality is pretty usable.